fix(contracts): increase size of members iterator variable
This bug was found by Veridise during their audit of Semaphore.

fix #201

+++ packages/contracts/contracts/Semaphore.sol
@@ -89,7 +89,7 @@ contract Semaphore is ISemaphore, SemaphoreGroups {
         override
         onlyGroupAdmin(groupId)
     {
-        for (uint8 i = 0; i < identityCommitments.length; ) {
+        for (uint256 i = 0; i < identityCommitments.length; ) {
             _addMember(groupId, identityCommitments[i]);
 
             unchecked {
